% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_daybeep.R
\name{add_daybeep}
\alias{add_daybeep}
\title{Add dayvar and beepvar to time series dataframes}
\usage{
add_daybeep(df, id, datetime)
}
\arguments{
\item{df}{dataframe with time series data.}

\item{id}{The name of the column that contains participant ID variable}

\item{datetime}{Column name with participant birthdate}
}
\value{
Your dataframe with columns numbering the dates (dayvar), times (beepvar), and consecutive beeps (beepconsec).
}
\description{
Add dayvar and beepvar to time series dataframes
}
