% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups.R
\name{dupsbydemos}
\alias{dupsbydemos}
\title{Check the range of responses}
\usage{
dupsbydemos(
  x,
  redcapid,
  id,
  birthdate,
  gender,
  height,
  weight,
  ethnicity,
  email
)
}
\arguments{
\item{x}{dataframe.}

\item{redcapid}{Autogenerated id number from REDCap, Ethica, etc}

\item{id}{Study id (e.g., participant initials, IE001, etc.)}

\item{birthdate}{Column name with participant birthdate}

\item{gender}{Column name with participant gender}

\item{height}{Column name with participant height}

\item{weight}{Column name with participant weight}

\item{ethnicity}{Column name with participant ethnicity}

\item{email}{Optional argument. Column name with participant email}
}
\value{
A message indicating how many duplicates were removed from your dataset and your dataset with the removed duplicates.
}
\description{
Check the range of responses
}
\examples{
dupsbydemos(fake_data, redcapid="id", id="si", birthdate = "birthdate", gender="demo_2", height="height", weight="weight", ethnicity = "ethnicity", email="demo_11")
}
