% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixheight.R
\name{fixheight}
\alias{fixheight}
\title{Wrangle height data}
\usage{
fixheight(df, var)
}
\arguments{
\item{df}{A dataframe.}

\item{var}{Column name containing messy height data.}
}
\value{
\itemize{
 \item Your dataframe containing a new column with fixed heights, named height.
 \item A new dataframe titled heightflag that contains rows with heights that should be investigated for plausible values, errors, etc.
}
}
\description{
Wrangle height data
}
\examples{
fixheight(dat, "height")
fixheight(fake_data, "demo_9")
}
