% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixweight.R
\name{fixweight}
\alias{fixweight}
\title{Wrangle weight data}
\usage{
fixweight(df, var)
}
\arguments{
\item{df}{A dataframe.}

\item{var}{Column name containing messy weight data.}
}
\value{
\itemize{
 \item Your dataframe containing a new column with fixed weights, named weight.
 \item A new dataframe titled weightflag that contains rows with weights that should be investigated for plausible values, errors, etc.
}
}
\description{
Wrangle weight data
}
\examples{
fixheight(dat, "weight")
fixheight(fake_data, "demo_10")
}
