% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_centfunction.R
\name{gather_cent}
\alias{gather_cent}
\title{Gather centrality estimates from contemporaneous and temporal}
\usage{
gather_cent(pcc_object, pdc_object)
}
\arguments{
\item{pcc_object}{object containing partial contemporaneous correlations}

\item{pdc_object}{object containing partial directed correlation}
}
\value{
a dataframe with centrality estimates from both pcc and pdc networks
}
\description{
Gather centrality estimates from contemporaneous and temporal
}
