% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_selfunction.R
\name{item_sel}
\alias{item_sel}
\title{Select variables based on means and sds}
\usage{
item_sel(df, num)
}
\arguments{
\item{df}{dataframe.}

\item{num}{Number of top means/sds to pull.}
}
\value{
A list of the variables with the highest means and highest standard deviations.
}
\description{
Select variables based on means and sds
}
