% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpss_dffunction.R
\name{kpss_df}
\alias{kpss_df}
\title{KPSS data frame function}
\usage{
kpss_df(df, vars = NULL)
}
\arguments{
\item{df}{dataframe}

\item{vars}{option argument to specify a list of which vars to test}
}
\value{
A dataframe containing the kpss test statistic, p value, and logical column of true or false where true indicates stationairyt met, and false means the stationarity assumption is potentially violated.
}
\description{
KPSS data frame function
}
